add one more test by putting everything in a router
