+++ epoque-core/src/test/kotlin/io/github/uharaqo/epoque/impl/CommandExecutorSpec.kt
@@ -7,7 +7,6 @@ import io.github.uharaqo.epoque.api.Version
 import io.github.uharaqo.epoque.api.VersionedEvent
 import io.github.uharaqo.epoque.impl.TestEnvironment.TestCommand.Create
 import io.kotest.assertions.arrow.core.rethrow
-import io.kotest.assertions.arrow.core.shouldBeRight
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 import io.mockk.coEvery
@@ -25,29 +24,18 @@ class CommandExecutorSpec : StringSpec(
       val dummyCommandExecutor = dummyCommandExecutor(eventWriter = eventWriter)
 
       // when
-      val out = dummyCommandExecutor.execute(dummyJournalKey.id, Create("Whatever"))
-
-      // then: returned == written
-      out shouldBeRight CommandOutput(slot.captured)
+      val output = dummyCommandExecutor.execute(dummyJournalKey.id, Create("Whatever")).rethrow()
 
       // then
-      out.rethrow().events shouldBe listOf(
-        VersionedEvent(Version(3), resourceCreatedEventType, serializedEvent1),
-        VersionedEvent(Version(4), resourceCreatedEventType, serializedEvent2),
+      output.events shouldBe slot.captured
+      output shouldBe CommandOutput(
+        listOf(
+          VersionedEvent(Version(3), resourceCreatedEventType, serializedEvent1),
+          VersionedEvent(Version(4), resourceCreatedEventType, serializedEvent2),
+        ),
       )
     }
   },
 ) {
-  companion object : TestEnvironment() {
-    fun dummyCommandExecutor(eventWriter: EventWriter) =
-      CommandExecutor(
-        dummyJournalKey.groupId,
-        dummyCommandHandler,
-        dummyEventCodecRegistry,
-        dummyEventHandlerExecutor,
-        dummyEventLoader,
-        eventWriter,
-        dummyTransactionStarter,
-      )
-  }
+  companion object : TestEnvironment()
 }
