+++ epoque-core/src/test/kotlin/io/github/uharaqo/epoque/impl/JournalSpec.kt
@@ -1,12 +1,22 @@
 package io.github.uharaqo.epoque.impl
 
+import arrow.core.right
 import io.github.uharaqo.epoque.api.CanAggregateEvents
+import io.github.uharaqo.epoque.api.CommandInput
+import io.github.uharaqo.epoque.api.CommandOutput
+import io.github.uharaqo.epoque.api.CommandProcessor
+import io.github.uharaqo.epoque.api.CommandType
+import io.github.uharaqo.epoque.api.EventWriter
 import io.github.uharaqo.epoque.api.Version
+import io.github.uharaqo.epoque.api.VersionedEvent
 import io.github.uharaqo.epoque.api.VersionedSummary
 import io.github.uharaqo.epoque.impl.TestEnvironment.TestSummary
 import io.kotest.assertions.arrow.core.rethrow
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
+import io.mockk.coEvery
+import io.mockk.mockk
+import io.mockk.slot
 
 class JournalSpec : StringSpec(
   {
@@ -22,6 +32,35 @@ class JournalSpec : StringSpec(
 
       // then
       result shouldBe VersionedSummary(Version(2), TestSummary.Default(dummyEvents))
+    }
+
+    "Basic components work together as expected" {
+      // given
+      val eventWriter = mockk<EventWriter>()
+      val slot = slot<List<VersionedEvent>>()
+      coEvery { eventWriter.write(any(), capture(slot), any()) } returns Unit.right()
+
+      val commandExecutor = dummyCommandExecutor(eventWriter)
+      val commandType = CommandType.of<TestEnvironment.TestCommand.Create>()
+      val commandCodec = dummyCommandCodecRegistry.find(commandType).rethrow()
+      val testProcessor = TypedCommandProcessor(commandCodec, commandExecutor)
+      val commandProcessors = mapOf(commandType to testProcessor)
+      val processor: CommandProcessor = DefaultCommandRouter(commandProcessors)
+
+      // when
+      val command = TestEnvironment.TestCommand.Create("Integration")
+      val serialized = commandCodec.serialize(command).rethrow()
+      val output =
+        processor.process(CommandInput(dummyJournalKey.id, commandType, serialized)).rethrow()
+
+      // then
+      output.events shouldBe slot.captured
+      output shouldBe CommandOutput(
+        listOf(
+          VersionedEvent(Version(3), resourceCreatedEventType, serializedEvent1),
+          VersionedEvent(Version(4), resourceCreatedEventType, serializedEvent2),
+        ),
+      )
     }
   },
 ) {
